Building on Windows and Linux.

+++ Source/tools/builder/buildProject.sh
@@ -22,25 +22,25 @@ else
 	LIBRARY_PATH="$(realpath ${BUILDER_FOLDER}/../../DFPSR)"
 	SOURCE_CODE="${BUILDER_FOLDER}/main.cpp ${BUILDER_FOLDER}/Machine.cpp ${BUILDER_FOLDER}/generator.cpp ${LIBRARY_PATH}/collection/collections.cpp ${LIBRARY_PATH}/api/fileAPI.cpp ${LIBRARY_PATH}/api/bufferAPI.cpp ${LIBRARY_PATH}/api/stringAPI.cpp ${LIBRARY_PATH}/base/SafePointer.cpp"
 	"${BUILDER_CPP_COMPILER_PATH}" -o "${BUILDER_EXECUTABLE}" ${SOURCE_CODE} -std=c++14
-	if [ $? -ne 0 ]
-	then
+	if [ $? -eq 0 ]; then
+		echo "Completed building the Builder build system."
+	else
 		echo "Failed building the Builder build system, which is needed to build your project!"
 		exit 1
 	fi
-	chmod +x "${BUILDER_EXECUTABLE}"
 fi
+chmod +x "${BUILDER_EXECUTABLE}"
 
 # Call the build system with a filename for the output script, which is later given execution permission and called.
 SCRIPT_PATH="/tmp/dfpsr_compile.sh"
 echo "Generating ${SCRIPT_PATH} from $1"
-rm "${SCRIPT_PATH}"
-"${BUILDER_EXECUTABLE}" $@ "ScriptPath=${SCRIPT_PATH}";
-if [ $? -ne 0 ]
-then
-	echo "Failed building the Builder build system, which is needed to build your project!"
-	exit 1
+if [ -e "${SCRIPT_PATH}" ]; then
+	rm "${SCRIPT_PATH}"
 fi
-echo "Giving execution permission to ${SCRIPT_PATH}"
-chmod +x "${SCRIPT_PATH}"
-echo "Running ${SCRIPT_PATH}"
-"${SCRIPT_PATH}"
+"${BUILDER_EXECUTABLE}" $@ "ScriptPath=${SCRIPT_PATH}";
+if [ -e "${SCRIPT_PATH}" ]; then
+	echo "Giving execution permission to ${SCRIPT_PATH}"
+	chmod +x "${SCRIPT_PATH}"
+	echo "Running ${SCRIPT_PATH}"
+	"${SCRIPT_PATH}"
+fi
